Moving theme-provider into providers directory in ui
