+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import {ThemeProvider} from "@/components/ui/theme-provider";
+import {ThemeProvider} from "@/components/ui/providers/theme-provider";
 
 const inter = Inter({ subsets: ['latin'] });
 import { Footer } from '@/components/Footer';
